fix migration for openstack clusters
Since !1714 (93ea1b4f), we rely on Terraform outputs to build the
YAML-based hosts file. For these outputs to exist, the Terraform stage
has to be run first.

This commit adds a call to the Terraform stage to the migration script
so the cluster repo is in a usable state afterwards.