Don't check for releasenotes for hotfixes against devel
Because hotfixes against devel are basically backports of hotfixes to
the latest release. As Changelogs are scoped to a specific release now
and we do not maintain a Changelog file on devel anymore, there are
neither changes on the Changelog file nor a releasenote file in hotfixes
against devel. We have to skip the job therefore (or add a useless
releasenote, but that's not the preferred solution).

Fixes-Up: abe11649c9d1bd0e749d7f4e1dbc816c0db1f66a