ch-k8s-lbaas: Implement removal tasks
Until now there was no automation to deconfigure ch-k8s-lbaas components
once rolled out. This commits implements the necessary steps to remove
the ch-k8s-lbaas components if it is not enabled.
To ensure ch-k8s-lbaas is cleaned up, we cannot skip the playbook if it
is disabled anymore. So we are basically favoring robustness over
runtime here.