Increase nf_conntrack settings on frontend nodes
As we observed that the default values are not sufficient, i.e. if
ch-k8s-lbaas is enabled and SNAT'ing on the gateways is disabled.
If the conntrack table is filled up, keepalived may gets prevented from
sending its VRRP packets which can cause cascading failovers.

The settings are done on frontend nodes in general and are not limited
to setups with gateways as it makes sense to increase these setting in
general on nodes running the frontend services.

Closes: #837