openstack: Fix configuration of root_disk_size
The existing code was not working as expected as an individually
configured "root_disk_size" for a node had no effect.
This commit fixes that such that "root_disk_size" for a node is used if
set and a fallback to the disk property of the configured flavor is used
if unset.

Co-authored-by: brunos <bruno.schubert@cloudandheat.com>